refactor: move LotteryNumber
