+++ src/main/java/step2/NotEnoughNumberException.java
@@ -1,0 +1,4 @@
+package step2;
+
+class NotEnoughNumberException extends IllegalArgumentException {
+}
