+++ src/test/java/step2/LotteryNumberTest.java
@@ -6,10 +6,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Set;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -44,36 +40,4 @@ public class LotteryNumberTest {
         //@formatter:on
     }
 
-    static class LotteryNumber {
-        public static LotteryNumber of(Integer... numbers) {
-            return new LotteryNumber(new HashSet<>(Arrays.asList(numbers)));
-        }
-
-        private final Set<Integer> numbers;
-
-        public LotteryNumber(Set<Integer> numbers) {
-            this.numbers = Collections.unmodifiableSet(numbers);
-            if (numbers.size() != 6) {
-                throw new NotEnoughNumberException();
-            }
-        }
-
-        public Set<Integer> getNumbers() {
-            return numbers;
-        }
-
-        public int getMatched(LotteryNumber other) {
-            Set<Integer> intersection = new HashSet<>(numbers);
-            intersection.retainAll(other.numbers);
-            return intersection.size();
-        }
-
-        @Override
-        public String toString() {
-            return numbers.toString();
-        }
-    }
-
-    private static class NotEnoughNumberException extends IllegalArgumentException {
-    }
 }
